+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.2.1
+
+- Add important ligatures (for better enharmonic cycling)
+- Make secondary lig/aux/sec declaration parsing cleaner & order-independent
+- Fix `accidentalsHash` mutating original `SymbolCode[]` array
+- Fix 0-cent accidental chain stepsize breaking tuning config parsing.
+- 
+- Better XenNote not found error log messages explaining how the parsing was attempted and how it failed.
+
 ## 0.2.0
 
 - Text accidentals (fingerings)
@@ -17,7 +26,6 @@
 - Implement transpose & enharmonic function
 - Renamed project to 'Xen Tuner'
 - Run plugin as docked window, add tuning cache for better performance
-- 
 
 ## 0.1.0